Unfold tinyurl link to BitTorrent
The tinyurl link does not work as expected, at least for my browser and NAS.

1. Clicking the link will result in an "Address wasn't understood" browser error, because the `https:` link 301 redirects to a `magnet:` link which most browsers don't understand.
  - `Prevented navigation to “https://tinyurl.com/gpt4all-lora-quantized” due to an unknown protocol.`
2. Copying the tinyurl link into Synology Download Station will result in an error. It will _not_ be automatically unfolded to a `magnet:` link.

I had to open the link in a browser window with the developer console Networking tab open, and grab the 301 redirect location from there. Only then did everything work as expected.

I understand the desire to track accesses to the magnet link but I think that, in this case, the headache is not worth it.